add weather_process_file
