+++ index.js
@@ -3,6 +3,7 @@ const server = require("express")();
 const line = require("@line/bot-sdk");
 const axios = require("axios");
 const areaModule = require("./src/area")
+const weatherModule = require("./src/weather")
 
 // パラメーターの設定(LINE)
 const line_config = {
@@ -65,9 +66,9 @@ server.post('/bot/webhook', line.middleware(line_config), (req, res, next) => {
             let week_weather = selectArea == null ? "東京の天気\n" : `${selectArea.name}の天気\n`
             for(i = 0; i < res.data.daily.length; i++) {
               if(i < res.data.daily.length - 1) {
-                week_weather += responseMessage(res.data.daily[i]) + "\n\n"
+                week_weather += weatherModule.responseMessage(res.data.daily[i]) + "\n\n"
               } else {
-                week_weather += responseMessage(res.data.daily[i])
+                week_weather += weatherModule.responseMessage(res.data.daily[i])
               }
             }
             events_processed.push(bot.replyMessage(event.replyToken, {
@@ -94,37 +95,3 @@ server.post('/bot/webhook', line.middleware(line_config), (req, res, next) => {
     }
   )
 });
-
-function responseMessage(daily_data) {
-  return `${unixtimeToDate(daily_data.dt)} ${getWeatherEmoji(daily_data.weather[0].main)}
-最高気温：${kelvinToCelsius(daily_data.temp.max)}度
-最低気温：${kelvinToCelsius(daily_data.temp.min)}度
-降水確率：${Math.floor(daily_data.pop * 100)}%`
-}
-
-// unix時間の変換
-function unixtimeToDate(unixtime) {
-  let date = new Date(unixtime * 1000)
-  return `${date.getFullYear()}年${date.getMonth()+1}月${date.getDate()}日`
-}
-
-// 気温の変換
-function kelvinToCelsius(kelvin) {
-  const kelvinDegree = 273.15
-  return Math.floor(kelvin - kelvinDegree)
-}
-
-// 天気の絵文字を取得
-function getWeatherEmoji(main) {
-  let result = ""
-  const emojiList = {
-    "Clear": "\uDBC0\uDCA9",
-    "Clouds": "\uDBC0\uDCAC",
-    "Rain": "\uDBC0\uDCAA",
-    "Snow": "\uDBC0\uDCAB"
-  }
-  if (emojiList[main] != null) {
-    result = emojiList[main]
-  }
-  return result
-}
